Improve Oauth Authorization Server fallback logic
I'm dealing with the [Atlassian MCP
Server](https://mcp.atlassian.com/v1/sse). This server hosts the
`/.well-known/oauth-authorization-sever` without the path aware support
With the current version of the SDK, this will never hit the case where
we try the wellknown path without the pathname because it fails cors
both times due to how this server is configured. This throws an error
instead of returning undefined and letting the fallthrough logic work.

This updates the logic so that we always fallback on using the well
known server path without the pathname.