fix: disallow null (infinite) requested TTL
The specification only allows the server to return a null TTL, it doesn't allow the client to request that.
This is intentional, as it forces clients to make a choice betwen a specific TTL or not caring what the TTL is, rather than defaulting to keeping tasks forever all the time.

Allowing the requested value to be null was an implementation oversight that diverged from the spec.